Consolidate json-schema sync into _schema.py
Moves ensure_json_schema plus its supporting constants out of _legacy.py
and into modules/helpers/_schema.py — where the other schema/hash
helpers already live (calculate_hash, load_previous_hashes, save_hashes,
_schema_files_present).

Moved from _legacy.py to _schema.py:
  - ensure_json_schema (~55 lines)
  - JSON_SCHEMA_SYNC_FILES (30-tuple of local/remote schema paths)
  - JSON_SCHEMA_DIR (config/.schema directory)
  - HASH_FILE (schema hash sidecar)
  - JSON_SCHEMA_REFRESH_TTL_SECONDS (env-configurable TTL)
  - _JSON_SCHEMA_LAST_REFRESH_AT (module-scoped refresh timestamp)

_schema.py now imports CONFIG_DIR, GITHUB_BASE_URL, and ALLOWED_EXTENSIONS
from _legacy.py (dropping the outbound _legacy -> _schema import for the
schema-specific constants).

External call sites (quickstart.py, modules/output.py, modules/persistence.py,
tests/conftest.py, tests/test_json_schema_sync_manifest.py) already
access these via helpers.foo() and helpers.JSON_SCHEMA_* — the star
re-export in helpers/__init__.py keeps them working unchanged.

Reduces _legacy.py from 374 to 276 lines (-98). All 1077 tests pass.
